# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "VibeCoding社区 C 端（帖子、评论、点赞、通知、上传）与 B 端（内容管理、用户管理、数据统计）接口文档",
        "title": "VibeCoding 社区 API",
        "termsOfService": "https://github.com/vibecoding/community",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/vibecoding/community"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/notifications": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-通知"
                ],
                "summary": "通知列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "C端-通知"
                ],
                "summary": "标记已读",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通知ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/posts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-帖子"
                ],
                "summary": "帖子列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "话题名",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-帖子"
                ],
                "summary": "发帖",
                "parameters": [
                    {
                        "description": "content, type, media_urls, topic_names",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_c.CreatePostReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vibecoding_community_internal_model.Post"
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-帖子"
                ],
                "summary": "帖子详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vibecoding_community_internal_model.Post"
                        }
                    }
                }
            }
        },
...
        "/api/v1/posts/{id}/like": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-互动"
                ],
                "summary": "点赞",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "C端-互动"
                ],
                "summary": "取消点赞",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/me/posts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "C端-帖子"
                ],
                "summary": "我的帖子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_vibecoding_community_internal_model.Post": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_vibecoding_community_internal_model.PostMedia"
                    }
                },
                "status": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.PostStatus"
                },
                "type": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.PostType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_vibecoding_community_internal_model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_vibecoding_community_internal_model.PostMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_vibecoding_community_internal_model.PostStatus": {
            "type": "string",
            "enum": [
                "normal",
                "hidden",
                "flagged"
            ],
            "x-enum-varnames": [
                "PostStatusNormal",
                "PostStatusHidden",
                "PostStatusFlagged"
            ]
        },
        "github_com_vibecoding_community_internal_model.PostType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "video"
            ],
            "x-enum-varnames": [
                "PostTypeText",
                "PostTypeImage",
                "PostTypeVideo"
            ]
        },
        "github_com_vibecoding_community_internal_model.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_c.CreatePostReq": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "media_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "topic_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "text, image, video",
                    "type": "string"
                }
            }
        },
...
    }
}